Remove lazy loading
