--- internal/core/discussion/author.go
@@ -1,7 +1,0 @@
-package discussion
-
-type Author struct {
-	ID        string
-	Name      string
-	AvatarURL string
-}
