+++ internal/core/discussion/ports.go
@@ -2,11 +2,7 @@ package discussion
 
 import "context"
 
-type CommentLoader interface {
+type CommentRepo interface {
 	GetCommentsBySubjectID(ctx context.Context, subjectID string) ([]*Comment, error)
 	GetAuthorByID(ctx context.Context, id string) (Author, error)
-}
-
-type CommentRepo interface {
-	CommentLoader
 }
